pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["albina_kazartseva <albinak74@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
djangorestframework = "^3.16.0"
pillow = "^11.3.0"
djangorestframework-simplejwt = "^5.5.1"
drf-yasg = "^1.21.10"
django-filter = "^25.1"
celery = "^5.5.3"
eventlet = "^0.40.2"
django-celery-beat = "^2.8.1"
django-cors-headers = "^4.7.0"
requests = "^2.32.4"
coverage = "^7.10.2"
redis = "^6.2.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
